refactor: Move RateLimiter to cli.progress_bars
It's only used by the raw progress bars now.
